image_generator/requirements-in.txt

# server
flask
flask_cors

# image librarts
pillow

# ai
diffusers
torch==2.4.0
transformers
accelerate